+++ tests-ios/PickingTest/AppDelegate.m
@@ -42,7 +42,7 @@
     //int brk = 1;
 }
 
-- (void)setupScene
+- (void)setUpScene
 {
     ICScene *scene = [ICScene scene];
     [scene setSize:CGSizeTokmVec3(self.hostViewController.view.bounds.size)];
@@ -87,7 +87,7 @@
     self.window.rootViewController = self.hostViewController;
     [self.window makeKeyAndVisible];
     
-    [self setupScene];
+    [self setUpScene];
 
     return YES;
 }